==> scripts/09_BAmounds.Rmd <==
---
title: "Bronze Age mounds - Are they different?"
author: "Adela Sobotkova"
date: "2022-11-17"
output: html_document
---

This script summarizes environmental information on the Bronze Age burial mounds in the Yambol Province, and plots their location and pinpoints other mounds within Yambol that share similar environmental properties.

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(sf)
library(raster)
library(plyr)
library(tidyverse)
```

## Load 2009-2022 mound data
We have a total of 1090 mounds, extinct and uncertain mound features in Yambol.
```{r load-mounds}
# Yambol mounds
Yam_mnds <- readRDS("../output_data/Yam_dd_mnds.rds") # Yam_mnds.rds dataset is the most recent deduplicated from 27/12/2022
Y_region <- st_read("../data/YamRegion.shp")
head(Yam_mnds)
Yam_mnds %>% group_by(Type) %>% tally()
```
## Extract excavated Bronze Age mounds

We now separate out from the Yambol survey dataset the mounds that were later excavated and yielded Bronze Age (BA) cultural material. We do this in two batches: 

First, we filter mounds on the basis of TRAP ID list supplied by Todor and second, we create spatial data for mounds excavated before TRAP survey.

### Excavated post-survey
```{r BA-from-TRAP}
# Filter out the BA mounds from the FAIMS data
bronze <- c(6009,8007,6003,8345,8346,8357,8502) # Toshko provided these TRAP numbers
# He also suspects 8351, 9257, but let's run without them for now.
BA <- Yam_mnds %>% 
  filter(TRAP %in% bronze)

# Add literature dimensions to mound 8345, which was excavated before survey 
BA$DiameterMax[4] <- 37.5
BA$HeightMax[4] <- 3.2

```
Originally Todor indicated 6010 among the BA mound IDs, but visual inspection of 6010 pictures did not indicate an excavated mound at this number. I tested mound 6011 instead of 6010, because the image for 6011 shows an excavated mound and I thought maybe an error happened, but its prominence was less than 6% as it sits on a flat ridge and Todor explained that the mound was half-excavated after survey (2012), for which there is a clear candidate in 6003 (Lozyanska - UTM 35 T 468101 4679119)

### Excavated before survey
The mounds excavated before survey had their attributes specified in the AKB reports and their data and coordinates follow below:

Drazhevo - Height: 2.7 m; Diameter: 37.0 m; Land: pasture on the rocky area
Straldzha - Height: 4.0 m; Diameter: 50.0 m; Land: wines and pasture; this corresponds to 9969 
Popovo - Height: 6.8 m; Diameter: 48.0 m; Land: wines and pasture
8345 Mogila village - Height: 3.2 m; Diameter: 37.6 m; Land: pasture on the rocky area 

```{r BA-excavated-Todor}
# Add the few mounds excavated before TRAP coverage >> these can also be pulled out of the AOR dataset
excavated_before <- data.frame(place = c("Drazhevo", 
                           "Straldzha",
                           "Popovo"),
                 Easting=c(455442,479522,479391),
                 Northing=c(4710935,4715730,4668279),
                 HeightMax=c(2.7,4.0,6.8),
                 DiameterMax = c(37,50,48),
                 LU_Around = c("Pasture","Perennial","Perennial"),
                 Type = c("Burial Mound","Burial Mound","Burial Mound"),
                 Source = c("Excavation","Excavation","Excavation"))
excavated_before <- st_as_sf(excavated_before, coords =c("Easting", "Northing"), crs = 32635)

```

Combine the excavated and survey data
```{r BA-combo}
# Combine both BA datasets and spatialize
BA_excavated <- rbind.fill(BA, excavated_before) # combine the attributes by row
BAcoords <- rbind(st_coordinates(BA),st_coordinates(excavated_before)) # combine coordinates
BA_excavated <- cbind(BA_excavated, BAcoords) # add coordinate column to the attributes
BA_excavated <- st_as_sf(BA_excavated, coords = c("X", "Y"), crs = 32635) # convert to a simple feature
```

## Map the mounds

Compare the excavated Bronze Age mound locations with the other Yambol mounds
```{r}
library(mapview)
mapview(BA_excavated) + mapview(Yam_mnds, cex = 2, zcol = "Type") 
#mapview(excavated_beforef) + mapview(Yam_mnds, cex = 2, zcol = "Type")
```

```{r eval=FALSE}

#tiff("../figures/Figure01.tiff", height = 1600, width = 1000)
plot(Y_region$geometry) 
     #main = "Excavated Bronze Age mounds (red circles) \n and other mound locations in Yambol")
plot(Yam_mnds$geometry, add = T, col = "lightgrey");
plot(BA_excavated$geometry, pch = 16, col = "red", add =T)
#dev.off()

```
```{r Bulgaria-overview}
library(leaflet)
leaflet(st_transform(Yam_mnds,4326)) %>% 
  addProviderTiles("Stamen.TonerBackground", group="Background") %>% 
  addProviderTiles("Stamen.TonerLite", group = "Lite") %>% 
  addProviderTiles("Stamen.Watercolor", group = "Watercolor") %>%
  ## add Mounds and BA mounds
  addCircles(color = "black", radius = 1) %>% 
  addCircleMarkers(data = st_transform(BA_excavated,4326), color = "red", radius = 5) %>% 
  addPolygons(data =st_transform(Y_region, 4326), color = "darkgrey") %>% 
#data=Yam_, group="Features", color = "black",
 #                  radius = 7,   fillOpacity = 0.75, stroke = FALSE,
                   #clusterOptions = markerClusterOptions(),
  #           popup = paste0("FeatureID: ", features$identifier,
addMiniMap(
    # tiles = providers$Esri.WorldTopoMap,
    tiles = providers$Stamen.TonerLite,
   # tiles = providers$Esri.OceanBasemap,
    position = 'topright', 
    width = 300, height = 300,
    zoomLevelOffset = -4,
    #centerFixed = c(),
    toggleDisplay = TRUE) %>%
  
addLayersControl(
    baseGroups = c("Lite","Watercolor","Background"),
    options = layersControlOptions(collapsed = T)) %>% 
  
addScaleBar(position = c("bottomleft"), 
            options = scaleBarOptions(maxWidth = 300, metric = TRUE,imperial = FALSE))

```

```{r Yambol-vis-Fig1}

ggplot(Y_region)+
  geom_sf(lwd = 1, fill = "white")+
  geom_sf(data = Yam_mnds, color = "grey")+
  geom_sf(data = BA_excavated, color = "red", cex = 2)+
  theme_minimal()

```


## Extract prominence and elevation data from raster
Location alone does not tell us much about the mounds. Let's extract additional values from the underlying 30m res ASTER imagery so we can compare the mounds' location formally.

We first load the ASTER layer and a derived slope and aspect raster.
```{r raster-load}
Y_elev <- raster("../output_data/large/Yelev32635.tif")
Y_aspslope <- brick("../output_data/large/Yaspslope.tif")
Y_tri <- brick("../output_data/large/YflowroughTRI.tif") # see script 00 line 380, TRI is the layer no.4 in the sequence of brick(Y_flow, Y_rough, Y_TPI, Y_TRI)
```

Next, we extract values from raster at mound locations. 

An  alternative to extraction is to, when grabbing the excavated mounds, get the information from the AOR reports, so as to have as complete a record of the morphology as possible. Here I enrich the BA mound data with environmental information extracted from ASTER raster. I simply overwrite the data where it exists so as to be consistent
```{r sample-elev-aspect-slope-22}
# Sample elevations at mound locations
BA_excavated$elevAster <- raster::extract(Y_elev, st_coordinates(BA_excavated))

# Extract values from slope/aspect rasters created in script 00
BA_excavated$slopeAster <-  raster::extract(Y_aspslope$Yaspslope.1, st_coordinates(BA_excavated))

BA_excavated$aspectAster <-  raster::extract(Y_aspslope$Yaspslope.2, st_coordinates(BA_excavated))

BA_excavated$TRI <- raster::extract(Y_tri$YflowroughTRI.4, st_coordinates(BA_excavated))

BA_excavated$TPI <- raster::extract(Y_tri$YflowroughTRI.3, st_coordinates(BA_excavated))

BA_excavated$rough <- raster::extract(Y_tri$YflowroughTRI.2, st_coordinates(BA_excavated))

# Calculate prominence 

library(FSA) # we need an additional library for the perc() function
     # check the fce is here
BA_excavated$prom250mbuff <- raster::extract(Y_elev,# raster containing elevation data
                        st_coordinates(BA_excavated), # centroids of mounds
                        buffer = 250, # actual buffer size in crs units, in this case 250m  or ca 22x22 cells around kernel
                        fun = function(x){perc(x,x[length(x)/2],"lt", na.rm = FALSE, digits = 2)})

```

## Plot BA mounds with the prominence
In the next chunk we view the relative visual control of the surrounding terrain (prominence measures the percentage of area visible from mound within 360 degrees and 250m radius around the mound), and the accessibility. TRI is topographic ruggedness representing another measure of relative elevation above the terrain (it is only present in survey mounds and needs to be calculated for all excavated mounds)
```{r mapview-prom-BA}
library(leafsync)

pr <- mapview(BA_excavated, zcol ="prom250mbuff")
tri <- mapview(BA_excavated, zcol = "TRI")
sync(pr, tri)
```
## Tabulate and summarize the Bronze Age mound attributes
Let us now look at some attributes of interest such as morphology, elevation, slope, etc.

```{r table-BA-prom}
library(formattable)

highlight <- color_tile("yellow","yellow")

formattable(BA_excavated %>% 
  st_drop_geometry() %>% 
  dplyr::select(TRAP, HeightMax, DiameterMax, elevAster, slopeAster, prom250mbuff, LU_Around, LU_Top, TPI, rough), list(
  area(col = 6, row = 3 ) ~ highlight,
  area(col = 6, row = 6 ) ~ highlight,
  area(col = 1, row = 3 ) ~ highlight,
  area(col = 1, row = 6 ) ~ highlight,
  area(col = 6, row = 9 ) ~ highlight,
  area(col = 1, row = 9) ~ highlight
  ))
```
We might as well summarize these values to get an idea of the range present in BA mounds
```{r summaries}
summary(BA_excavated[,c(6:7,12:18)])
```

An average Bronze Age mound has the height of 3.5 m, diameter around 35 m, the prominence mean of 70%, TRI is around 1.6 and roughness at 5.  We may be able to extend these findings to the other Yambol mounds.

Most mounds are located in pastures(5) or rock outcrops with sparse grass cover, some scrub(2) and occasionally vineyards(2) .
```{r landuse}
BA_excavated %>% 
  st_drop_geometry() %>% 
  group_by(LU_Around) %>% 
  tally() %>% 
  arrange(desc(n))

BA_excavated %>% 
  st_drop_geometry() %>% 
  group_by(LU_Top) %>% 
  tally() %>% 
  arrange(desc(n))

```
As for the previous summary of attributes, the means are always deceptive, and small errors might cause big deviations (earlier 6010 was conflated with 6003). The prominence mean, for example, might be higher if the two low-prominence mounds were not included (6010 and 8351)


```{r scatterplot}
plot(BA_excavated %>%
       st_drop_geometry() %>% 
       dplyr::select(HeightMax, DiameterMax, elevAster, slopeAster, prom250mbuff))

plot(Yam_mnds %>%
       st_drop_geometry() %>% 
       filter(Type == "Burial Mound") %>% 
       dplyr::select(HeightMax, DiameterMax, elevAster, slopeAster, prom250mbuff))
```


Most of the mounds are in areas of high natural prominence, one exception confirms the rule: Straldzha mound is in a largely flat area of a Pleistocene lake with prominence below 50%. The low score may be caused by the coarse resolution of the 30m ASTER image, the 250m radius (disadvantages mounds located on elevated platforms). The other option is that the mound-builders cared less about elevated location, and sought dissociation from mounds on outcrops (already standing), or hoped to imbue the mound with prominence by raising its height.

## Predict mounds on the basis of prominence and TRI
The mean and median prominence values for BA mounds can provide a guide for other potentially Bronze Age mounds (if we choose a single factor)
The following map will show the mounds within the Yambol dataset that have prominence higher than the mean of 70% (265 features) and whose topographic ruggedness is over the mean of 1.6 (52). The lighter the point color the higher their prominence and the more probable their Bronze Age origin 
```{r mapview-comp-I}
library(mapview)
library(leafsync)

potentialBA <- Yam_mnds %>% 
  dplyr::filter(prom250mbuff > 70 & Type == "Burial Mound" & TRI >1.6) %>% 
  mapview(zcol = "prom250mbuff")

prom70 <- Yam_mnds %>% 
  filter(prom250mbuff > 70 & Type == "Burial Mound") %>% 
  mapview(zcol = "prom250mbuff")

BA <- mapview(BA_excavated)

sync( BA, potentialBA,prom70,ncol=3)
```

## With height
Setting the height to 3+ m produces an interesting result.  Over 3 m high and 70%+ prominence results in 52 features, same as without the height constraint, but they are in different locations from the 52 features selected on the basis of prominence and TRI. Compare the former (tri-based) features with height-based ones in pane 2 and 3.
```{r mapview-comparison-II}
library(mapview)
library(leafsync)

potentialBA3m <- Yam_mnds %>% 
  filter(HeightMax > 3 & 
        # elevAster > 170 &
        # DiameterMax > 30 &&
        # LU_Around == "Scrub" | LU_Around == "Pasture" &&
        #   LU_Top == "Pasture" | LU_Top == "Scrub" &&
         prom250mbuff > 70 &
        # TRI > 1.6 &&
         Type == "Burial Mound" ) %>% 
          mapview()
           
BA <- mapview(BA_excavated)
sync(BA,potentialBA,potentialBA3m, ncol=3)

```
### Tmap Merged Potential BA mounds (3m and TRI >1.6)
```{r}
library(tmap)

potentialBA <- Yam_mnds %>% 
  filter(TRI > 1.6 & 
         prom250mbuff > 70 &
         Type == "Burial Mound" )

potentialBA3m <- Yam_mnds %>% 
  filter(HeightMax > 3 & 
         prom250mbuff > 70 &
         Type == "Burial Mound" )

tm_shape(Yam_mnds, bbox = st_bbox(Y_region))+
  tm_symbols(shape = 2, col="#666666", size = 0.02) +
tm_shape(Y_region) +
 tm_borders(col = "grey", 
              lwd = 2)+
tm_shape(potentialBA, bbox = st_bbox(Y_region)) + 
  # tm_polygons(col="cornsilk2", border.alpha = 0)+
  # tm_shape(features) + 
  tm_symbols(shape = 21, col="red", 
            # size = 0.1,
             alpha = 0.5) +  #6 and 26 are downward triangles
tm_shape(potentialBA3m, bbox = st_bbox(Y_region)) + 
  # tm_polygons(col="cornsilk2", border.alpha = 0)+
  # tm_shape(features) + 
  tm_symbols(shape = 21, col="yellow", 
             #size = 0.1,
             alpha = 0.5) +  #6 and 26 are downward triangles
  tm_compass(type = "arrow", 
             position = c("left", "bottom")) +
  tm_scale_bar(position = c("left", "bottom"),
               breaks = c(0,5,10), 
               text.size = 0.75) + 
  #tm_credits("Sobotkova 2020", position = c("right", "bottom"), size = 0.8) +
  #tm_layout(title = "Registered standing features") 
  tm_layout(title = "A")
```
### Tmap Merged Potential BA mounds (3m and TRI >1.6)
```{r}
library(tmap)

potentialBA <- Yam_mnds %>% 
  filter(TRI > 1.6 & 
         prom250mbuff > 70 &
         Type == "Burial Mound" )

potentialBA3m <- Yam_mnds %>% 
  filter(HeightMax > 3 & 
         prom250mbuff > 70 &
         Type == "Burial Mound" )

tm_3m <- tm_shape(Yam_mnds, bbox = st_bbox(Y_region))+
  tm_symbols(shape = 2, col="#666666", size = 0.02) +
tm_shape(Y_region) +
 tm_borders(col = "grey", 
              lwd = 2)+
tm_shape(potentialBA, bbox = st_bbox(Y_region)) + 
  tm_symbols(shape = 21, col="red",  #6 and 26 are downward triangles
            # size = 0.1,
             alpha = 0.5)+
  tm_shape(BA_excavated) + 
  tm_symbols(shape = 24, col="black",  #6 and 26 are downward triangles
             size = 0.3)+
 tm_layout(scale =2, title = "A")

tm_tri <- tm_shape(Yam_mnds, bbox = st_bbox(Y_region))+
  tm_symbols(shape = 2, col="#666666", size = 0.02) +
 tm_shape(Y_region) +
    tm_borders(col = "grey", 
              lwd = 2)+
 tm_shape(potentialBA3m, bbox = st_bbox(Y_region)) + 
  tm_symbols(shape = 21, col="yellow", 
             #size = 0.1,
             alpha = 0.5) +  #6 and 26 are downward triangles
 tm_shape(BA_excavated) + 
  tm_symbols(shape = 24, col="black",  #6 and 26 are downward triangles
             size = 0.3)+
  tm_compass(type = "arrow", 
             position = c("left", "bottom")) +
  tm_scale_bar(position = c("left", "bottom"),
               breaks = c(0,5,10), 
               text.size = 0.75) + 
  #tm_credits("Sobotkova 2020", position = c("right", "bottom"), size = 0.8) +
  #tm_layout(title = "Registered standing features") 
  tm_layout(scale =2, title = "B")

tmap_arrange(tm_3m, tm_tri)

```
```{r print, eval=FALSE}
tiff("../figures/Figure04.tiff", width = 1600, height = 1000)
tmap_arrange(tm_3m, tm_tri)
dev.off()
```


## Table the potential BA mounds

```{r BA-3m}
potentialBA3m <- Yam_mnds %>% 
              st_drop_geometry() %>% 
  filter(HeightMax > 3 & 
           prom250mbuff > 70 &
         Type == "Burial Mound" )
#write.csv(potentialBA3m, "../output_data/potentialBA3m.csv")

formattable(potentialBA3m)

```

```{r BA-TRI}
potentialBA <- Yam_mnds %>% 
              st_drop_geometry() %>% 
  filter(TRI > 1.6 & 
           prom250mbuff > 70 &
         Type == "Burial Mound" )
#write.csv(potentialBA, "../output_data/potentialBA.csv")

formattable(potentialBA)
```

## Height and Elevation comparisons between BA and Yambol mounds

### Height Boxplots
```{r}
hist(BA_excavated$elevAster)
boxplot(BA_excavated$HeightMax[BA_excavated$HeightMax>0], Yam_mnds$HeightMax[Yam_mnds$HeightMax>0], names = c("BA mounds", "all mounds"), main = "Maximum mound height")

```

### Elevations
```{r }
#tiff(filename="../figures/Figure02.tiff", width=600, height=900, units = "mm", res = 300)
Yam_mnds %>% 
  ggplot() +
  geom_histogram(aes(x=elevAster,  fill = "All"), binwidth = 10)  +
  geom_histogram(data = BA_excavated, aes(x=elevAster,  fill = "Bronze Age"),  binwidth = 10) +
  labs(title = "Mound elevations in Yambol",
       x = "Elevation (m)",
       y = "Count")+
  theme_bw(base_size=60)+
  scale_fill_grey( name = "Mounds") + 
  theme(legend.position = "bottom") 
        #legend.text = element_text(size=48),
        #axis.text=element_text(size=48),
        #axis.title=element_text(size=48,face="bold"))
#dev.off()#
```

### Prominence comparison
```{r}
Yam_mnds %>% 
  ggplot() +
  geom_histogram(aes(x=prom250mbuff), binwidth = 2)  +
  geom_histogram(data = BA_excavated, aes(x=prom250mbuff), fill = "yellow", binwidth = 2) +
  labs(title = "Prominence of Yambol relief and BA mounds (in yellow)",
       x = "Prominence (%)",
       y = "Count")+
  theme_bw()

```



```{r}
BA_excavated %>% 
  ggplot(aes(x = LU_Around, y = HeightMax)) + 
  geom_violin(alpha = 0)+
  geom_jitter(color = "tomato")+
  theme_bw()+
  theme(panel.grid.major = element_blank())+
  labs(x = "Feature type",
       y = "Height of feature",
       title = "Mound height by type")
```
### Prominence densities

To get prominence plotted as densities rather than histograms, one must rerun ProminenceMC script.
```{r MC-densities, eval = FALSE}

# Calculate the mound densities

library(foreach)
library(purrrlyr)

#note the exposition pipe operator $, which works as dataframe$variable
BAmounds_densities <- BA_excavated %$%  
  prom250mbuff %>%
  density(from = 0,
            to = 100,
            n = 1201) %>% 
   broom::tidy() %>%
   tibble::as_tibble() %>%
  dplyr::mutate(y = y * 1201) %>%
  dplyr::rename(Prominence = x,
                Frequency = y)

ggplot() +
  geom_line(data = yambol_region_densities,
            mapping = aes(x = Prominence,
                          y = Frequency),
            col = "lightgrey") +
  geom_ribbon(data = yambol_densities_wide,
              mapping = aes(x = Prominence,
                            ymin = `Lower CI`,
                            ymax = `Upper CI`),
              col = "darkgrey",
              alpha = 0.5) +
  geom_line(data = mounds_densities,
               mapping = aes(x = Prominence,
                             y = Frequency),
               color = "red", size = 2)+
  geom_line(data = BAmounds_densities,
               mapping = aes(x = Prominence,
                             y = Frequency),
               color = "pink", size = 2)+
  theme_bw()+
  labs(x = "Topographic Prominence (%)")
```

```{r eval = FALSE, echo = FALSE}
tiff("../figures/Figure03.tiff", width = 2000, height = 1600)
ggplot() +
  geom_line(data = yambol_region_densities,
            mapping = aes(x = Prominence,
                          y = Frequency),
            col = "lightgrey") +
  geom_ribbon(data = yambol_densities_wide,
              mapping = aes(x = Prominence,
                            ymin = `Lower CI`,
                            ymax = `Upper CI`),
              col = "darkgrey",
              alpha = 0.5) +
  geom_line(data = mounds_densities,
               mapping = aes(x = Prominence,
                             y = Frequency),
               color = "red", size = 6)+
  geom_line(data = BAmounds_densities,
               mapping = aes(x = Prominence,
                             y = Frequency),
               color = "pink", size = 6)+
  theme_bw(base_size = 60)+
  labs(x = "Topographic Prominence (%)")
dev.off()
```
