==> scripts/10_Bunkers.Rmd <==
---
title: "Bunkers"
author: "Adela Sobotkova"
date: "2022-11-12"
output: html_document
---
```{r setup, include=FALSE}
library(knitr)
library(rmdformats)
library(formatR)

## Global options
options(max.print="75")
opts_chunk$set(echo=TRUE,
	             cache=TRUE,
               prompt=FALSE,
               tidy=TRUE,
               comment=NA,
               message=FALSE,
               warning=FALSE)
opts_knit$set(width=75)
```

# Explore and visualize military installations on the Cold-War border

Yambol region borders on Turkey and Greece and with the Tundzha River flowing through the border in a deep ravine overgrown with forests, this region represented a prime ground for birectional movement. Movement of populations is posited here since the first farming immigrants in the 6th millennium. As the arid conditions of the early Holocene were letting up and humidity allowed for higher vegetation to spread, the first farmers settled near the Tundzha River, planting crops and building the first permanent settlements in the area. 8000 years later, this was an administrative border between two hostile worlds, carefully guarded and with restricted movement. Military forces on both sides set up observation posts and defensive locations so as to keep control over the territory. 

This scripts visualizes the locations of military installations recorded in place of burial mounds (symbols in Soviet topographic maps).

This script also processes Angel's digitized Map Mounds - this section needs to eventually be moved to a separate script!!

To DOES: 
 - national boundaries for a good inset map
 
```{r libraries}
library(tidyverse)
library(sf)
library(raster)
```

## Load Data

### 2009-2022 FAIMS mound data, map mounds, and 2018 - 2022 military feature data
```{r get-mnd-data}
getwd()
#Yam_mnds <- readRDS("../output_data/features_dd_later_27Dec.rds")

# Clipped deduplicated 1073 mound data within the borders of Yambol
Yam_mnds <- read_rds("../../MoundMerging2023/output_data/mounds_dd_Yam.rds")

Yam_mil <- read_csv("../output_data/MIlitaryFeatures-OR12Dec22.csv")
mapmnds <- read_sf("../data/MapMounds32635.shp")

```
### Add and clean map mounds Angel digitized in June 2024
this concerns mounds north of Yambol within 25 km buffer 
```{r}
mapmnds_a <- read_sf("../data/angel_mapmnds/MoundsAngel.shp")


# fix angels' mounds IDs and Symbols
mapview(mapmnds_a, cex = 3) + mapview(mapmnds, cex =1)
mapmnds_a$Id <- 400000:400284
mapmnds_a$createdBy <- "Angel Grigorov"
mapmnds_a$createdAt <- "2024-06-06"
mapmnds_a <- mapmnds_a %>% 
  st_transform(crs = 4326) %>% 
  mutate(Latitude = st_coordinates(.)[,2],
         Longitude = st_coordinates(.)[,1])%>% 
   st_transform(crs = 32635) 

# this assignment does not work
mapmnds_a <- mapmnds_a %>% 
  mutate(MapSymbol = case_when(is.na(MpSymbol) ~ "Sunburst",
                               !is.na(MpSymbol) ~ MpSymbol)) %>%
  dplyr::select(-MpSymbol) %>% 
  mutate(Notes = "digitized in 2024")
#st_write(mapmnds_a, "../data/MapMoundsA32635.geojson")


# merge Angel's with main mapped mounds
mapmnds <- mapmnds %>% 
  rename(Id =identfr, createdBy=cretdBy, createdAt = FtrTmst,Latitude = Latitud, Longitude = Longitd, MapSymbol = MpSymbl, Notes = Note) %>% 
  dplyr::select(Id, createdBy, createdAt, Latitude, Longitude, MapSymbol, Notes) %>% 
  rbind(mapmnds_a)

```

### Elevation raster and Yambol region boundary
```{r get-elev}
Y_region <- st_read("../data/YamRegion.shp")

Y_elev <- raster("../output_data/large/Yelev32635.tif")
plot(Yam_mnds$geometry, cex = sqrt(as.numeric(Yam_mnds$HeightMax)), main = " Yambol region burial mound distribution")
```

### Explore the military features a bit

Most seem collected in 2018 and 2022 seasons, even though we have seen emplacements since 2009 (on Boyanovo quarry south slope)
```{r explore-mil}
tail(Yam_mil)
Yam_mil <- Yam_mil[-79,]

unique(Yam_mil$FacingDirection)
unique(Yam_mil$CentralFeature)
#Yam_mil$feature.facing <- factor(Yam_mil$feature.facing)

Yam_mil %>% 
  group_by(CentralFeature) %>% 
  tally() %>% 
  arrange(desc(n))

# check
library(lubridate)
Yam_mil %>% 
  left_join(Yam_mnds, by=c("MoundID"="TRAP")) %>% 
  dplyr::select(MoundID, Date, Type.x) %>% 
  mutate(Year = year(Date)) %>% 
  group_by(Year) %>% tally()

```

## Plot map mounds 
are good for the background, but now the file contains all kinds of benchmarks. Let's filter to the sunburst, brown sunburst-like symbol, which in 99% cases leads to a burial mound.

```{r get-map-mnd}
table(mapmnds$MapSymbol)  # 'circle' seems the best think to search on with ca 4289records

mapmnds <- mapmnds %>% 
  filter(grepl("circle|Sunburst",MapSymbol))

plot(mapmnds$geometry, main = "Mound symbols in Soviet maps"); plot(Y_region$geometry,border = "red",add = T)
```


## Create bunkers layer

and streamline their types for easier plotting
```{r bunkers}
Yam_bunkers <- Yam_mnds %>% 
  dplyr::filter(TRAP %in% Yam_mil$MoundID) %>% 
  full_join(Yam_mil, by = c("TRAP"="MoundID"))  # can be done via left join from Yam_mil layer

# we need a factor to visualize military feature type
Yam_bunkers$CentralFeature <- factor(Yam_bunkers$CentralFeature)

# consider simplifying Tank and Vehicle emplacement into one category
levels(Yam_bunkers$CentralFeature) <- c("Bunker", "Infantry trench","Tank emplacement","Vehicle emplacement")


# rename CentralFeature to Feature type for better legends
names(Yam_bunkers)[33] <- "Feature type"

st_write(Yam_bunkers, "../output_data/Yam_bunkers.shp", append = FALSE)
```

## Plot bunker types with mapview
Mapview is great for interactive viewing in the digital document.
```{r plot-bunkers}
library(mapview)

mapview(Yam_mnds, cex = 1) + mapview(Yam_bunkers, zcol = "Feature type")

```   
## Great, let's explore different plotting packages

circle = bunkers, 19, 20, 21 - color fill
square = trench, 7, 22
u/horseshoe = emplacement, 18, 25
mund triangle - 2,17,24

Maps: 
1: Yambol with all mounds with an inset of BG (no topography, minimalist graphics)
2: Border region with mounds and military installations visually differentiated
3: Typology of mil.installations
4: Sitovo site with different installations?

### Yambol with an inset of BG - Leaflet

```{r leaflet}
library(leaflet)
library(leafem)
leaflet() %>% 
  addProviderTiles("Esri.WorldImagery") %>% 
  addCircleMarkers(data = st_transform(Yam_bunkers, 4326)) %>% 
  addMiniMap(toggleDisplay = TRUE,
             position = "bottomright")

```


## Mounds and military features in tmaps
Let's create a printable overview Tmap differentiating between Yambol mounds/features and military installations


## Overview
```{r tmap-mapmnd-mil}
library(tmap)

# all the features throughout the region
overview_mapmnds <- tm_shape(mapmnds, bbox = st_bbox(Y_region))+
  tm_symbols(shape = 2, 
             col="#666666", 
             alpha = 0.5,
             size = 0.01) +
# regional boundary
tm_shape(Y_region) +
 tm_borders(col = "grey", 
              lwd = 2)+
# military installations
tm_shape(Yam_bunkers, bbox = st_bbox(Y_region)) + 
  # tm_polygons(col="cornsilk2", border.alpha = 0)+
  # tm_shape(features) + 
  tm_symbols(shape = 25, 
             col="Feature type", 
             palette = "plasma", # try Set2 ,Set3, viridis or other palettes
          #   title = "Feature type",
             size = 0.2) +  #6 and 25 are downward triangles
  tm_compass(type = "arrow", 
             position = c("right", "top")) +
  tm_scale_bar(position = c("right", "top"),
               breaks = c(0,5,10), 
               text.size = 0.75) + 
  #tm_credits("Sobotkova 2020", position = c("right", "bottom"), size = 0.8) +
  #tm_layout(title = "Registered standing features") 
  tm_layout(title = "A",
            legend.position = c("left", "bottom"))

overview_mapmnds

```
```{r tmap-yam-mm}

overview_yammnds <- tm_shape(Yam_mnds, bbox = st_bbox(Y_region))+
  tm_symbols(shape = 2, 
             col="#666666", 
             alpha = 0.5,
             size = 0.01) +
# regional boundary
tm_shape(Y_region) +
 tm_borders(col = "grey", 
              lwd = 2)+
# military installations
tm_shape(Yam_bunkers, bbox = st_bbox(Y_region)) + 
  # tm_polygons(col="cornsilk2", border.alpha = 0)+
  # tm_shape(features) + 
  tm_symbols(shape = 25, 
             col="Feature type", 
             palette = "plasma", # try Set2 ,Set3, viridis or other palettes
          #   title = "Feature type",
             size = 0.2) +  #6 and 25 are downward triangles
  tm_compass(type = "arrow", 
             position = c("right", "top")) +
  tm_scale_bar(position = c("right", "top"),
               breaks = c(0,5,10), 
               text.size = 0.75) + 
  #tm_credits("Sobotkova 2020", position = c("right", "bottom"), size = 0.8) +
  #tm_layout(title = "Registered standing features") 
  tm_layout(title = "A",
            legend.position = c("left", "bottom"))

overview_yammnds

#pdf(overview_yammnds, "figures/figure2.tiff")

```


## Military feature typology - detailed views
We need maps zooming in to differentiate the different locations of vehicle emplacements versus bunkers/observation towers etc.
```{r tmap-detail}
library(tmap)
tmap_mode("view")
tm_shape(Yam_bunkers, bbox = st_bbox(st_buffer(st_geometry(Yam_bunkers[1,]), 20000))) + 
  tm_dots(shape = "Feature type", #6 and 26 are downward triangles
            alpha= 0.4,
            title = "Type of \nmilitary features",
             size = "HeightMax", # was 0.3
          col = "yellow") +
   tm_text(text = "TRAP", size = 1)+  
tm_shape(mapmnds) +
  tm_dots(size = 0.01)+
tm_shape(Y_region) +
 tm_borders(col = "grey",
              lwd = 2)+
  tm_compass(type = "arrow", 
             position = c("left", "top")) +
  tm_scale_bar(position = c("left", "top"),
               breaks = c(0,5,10), 
               text.size = 0.75)  +
  tm_layout(title = "B", legend.outside = TRUE) +

  tm_minimap()


```
### Export data for Angel's viewsheds
```{r}
# filter bunkers
features %>% 
  filter(TRAP %in% c(8595, 9845, 8619,8515, 8707, 8535, 8244, 8271, 8798, 8230, 
                     9836, 9841, 9721, 8531, 8709,8711, 8278, 8798)) %>% # pull(TRAP) %>% sort()
  st_write("D:/Daaaaaataaaa/bunkers.shp")

group_by(Type) %>% 
  tally()
st_write(Yam_mnds, "../data/Yam_mnds.shp")
st_write(features, "../data/Yam_features.shp")
```


## Facing directions 

The direction the vehicle emplacements and other features are pointing can tell us about their role: offensive or defensive?

```{r directions}
table(Yam_bunkers$FacingDirection)
Yam_bunkers <- Yam_bunkers %>% 
  #st_drop_geometry() %>% 
  mutate(FD = case_when(
    FacingDirection == "Southeast" ~ "South", 
    FacingDirection == "Southwest" ~ "South", 
    FacingDirection == "Northeast" ~ "North",
    FacingDirection == "South" ~ "South",
    FacingDirection == "North" ~ "North",
    FacingDirection == "East" ~ "East",
    FacingDirection == "West"~ "West"))

table(Yam_bunkers$Feature_Facing)
```


### Explore tmap palettes
```{r eval=FALSE}
# You may need to install the dependency
#install.packages("shinyjs")
library(shinyjs)
tmaptools::palette_explorer()

```



### Combine feature type and facing direction at 30 km scale

```{r type-direction}
table(Yam_bunkers$FacingDirection)
table(Yam_bunkers$FD)

color_type <- rainbow(length(unique(Yam_bunkers$FD)))
library(tmap)

FeatureType_col<- tm_shape(Y_elev,bbox = st_bbox(st_buffer(st_geometry(Yam_bunkers[1,]),20000)))+
tm_raster(alpha = 0.6,
          title = "Elevation (masl)") +
tm_shape(Yam_bunkers, bbox = st_bbox(st_buffer(st_geometry(Yam_bunkers[1,]),20000))) + 
  tm_dots( col = "FD",
           shape = "Feature type",
          palette = color_type,
          title = "Facing \ndirection",
          alpha= 0.6,
             size = 0.3) +  
tm_shape(Y_region) +
 tm_borders(col = "grey",
              lwd = 3)+
  tm_compass(type = "arrow", 
             position = c("left", "top")) +
  tm_scale_bar(position = c("left", "top"),
               breaks = c(0,5,10), 
               text.size = 0.75)  +
  tm_layout(title = "B", legend.outside = TRUE)


FeatureType_bw <- tm_shape(Y_elev,bbox = st_bbox(st_buffer(st_geometry(Yam_bunkers[1,]),20000)))+
tm_raster(alpha = 0.6,
          palette = "Greys",
          title = "Elevation (masl)") +
tm_shape(Yam_bunkers, bbox = st_bbox(st_buffer(st_geometry(Yam_bunkers[1,]),20000))) + 
  tm_dots( col = "FD",
           shape = "Feature type",
          palette = color_type,
          title = "Facing \ndirection",
          alpha= 0.6,
             size = 0.3) +  
tm_shape(Y_region) +
 tm_borders(col = "grey",
              lwd = 3)+
  tm_compass(type = "arrow", 
             position = c("left", "top")) +
  tm_scale_bar(position = c("left", "top"),
               breaks = c(0,5,10), 
               text.size = 0.75)  +
  tm_layout(title = "B", legend.outside = TRUE)

FeatureType_bw
FeatureType_col

# Eventually we chose one, but let's compare side-by-side
current.mode <- tmap_mode("plot")
tmap_arrange(FeatureType_bw, FeatureType_col)
tmap_mode(current.mode)
```

### Combine feature type and facing direction at 10 km scale

```{r type-dir-10}
table(Yam_bunkers$FacingDirection)
table(Yam_bunkers$FD)

color_type <- rainbow(length(unique(Yam_bunkers$FD)))
library(tmap)
tm_shape(Y_elev,bbox = st_bbox(st_buffer(st_geometry(Yam_bunkers[1,]), 10000)))+
tm_raster(alpha = 0.6,
          palette = "Greys",
          title = "Elevation (masl)") +
tm_shape(Yam_bunkers, bbox = st_bbox(st_buffer(st_geometry(Yam_bunkers[1,]), 10000))) + 
  tm_dots( col = "FD",
           shape = "Feature type",
          palette = color_type,
          title = "Facing \ndirection",
          alpha= 0.6,
             size = 0.3) +  
tm_shape(Y_region) +
 tm_borders(col = "grey",
              lwd = 2)+
  tm_compass(type = "arrow", 
             position = c("left", "bottom")) +
  tm_scale_bar(position = c("left", "bottom"),
               breaks = c(0,5,10), 
               text.size = 0.75)  +
  tm_layout(title = "B", legend.outside = TRUE)
```


### Combine feature type and facing direction at 1 km scale around Sitovo[3] or [44]

[3] and [44] are point within the two major clusters in the border area
```{r type-dir-1}
table(Yam_bunkers$FacingDirection)
table(Yam_bunkers$FD)

color_type <- rainbow(length(unique(Yam_bunkers$FD)))
library(tmap)

box <- st_bbox(st_buffer(st_geometry(Yam_bunkers[44,]), 1000))
diff <-  (box[4]-box[2])/2
box[2] <- box[2]+diff/2
box[4] <- box[4]-diff/2
tm_shape(Yam_bunkers,bbox = box) + 
  tm_dots( col = "FD",
           shape = "Feature type",
          palette = color_type,
          title = "Facing direction",
          alpha= 0.6,
             size = 0.3) +  
tm_shape(Y_region) +
 tm_borders(col = "grey",
              lwd = 2)+
  tm_compass(type = "arrow", 
             position = c("left", "bottom")) +
  tm_scale_bar(position = c("left", "bottom"),
               breaks = c(0,0.5,1), 
               text.size = 0.75)  +
  tm_layout(title = "Sitovo Area", legend.outside = TRUE)
```


### Try facets on bearing - not great
```{r direction-facet}
tm_shape(Yam_bunkers) + 
  tm_dots()+
  tm_facets(by = "FD",free.coords = FALSE)+
  tm_shape(Y_region) +
 tm_borders(col = "grey",
              lwd = 2) +
  tm_style("cobalt")

```
### Try facets on types - not great either
```{r type-facets}
tm_shape(Yam_bunkers) + 
  tm_dots()+
  tm_facets(by = "Feature type",free.coords = FALSE)+
  tm_shape(Y_region) +
 tm_borders(col = "grey",
              lwd = 2) +
  tm_style("white")

```


### Bulgaria inset preparation
```{r inset-prep}
# Get admin boundaries for Bulgaria and neighbors and simplify by units of CRS (meters)
BG <- getData('GADM', country = 'BG', level = 0)
BG <- st_transform(st_as_sf(BG), 32635) %>% 
  st_simplify(dTolerance = 1000)

TR <- getData('GADM', country = 'TR', level = 0)
TR <- st_transform(st_as_sf(TR), 32635) %>% 
  st_simplify(dTolerance = 1000)

GR <- read_sf("../data/GR_bsm_admn_adm0_py_EuroGeoGraphics_2015_pp.shp")
GR <- st_transform(GR, 32635) %>% 
  st_simplify(dTolerance = 1000)
RO <- getData('GADM', country = 'RO', level = 0)
RO<- st_transform(st_as_sf(RO), 32635) %>% 
  st_simplify(dTolerance = 1000)
SE <- getData('GADM', country = 'RS', level = 0)
SE <- st_transform(st_as_sf(SE), 32635) %>% 
  st_simplify(dTolerance = 1000)
MK <- getData('GADM', country = 'MK', level = 0)
MK <- st_transform(st_as_sf(MK), 32635) %>% 
  st_simplify(dTolerance = 1000)
```
```{r quickplot}
# Plot - but first create a bounding box
box <- st_as_sfc(st_bbox(st_buffer(BG$geometry, 100000)))
# or extent

#e <- drawExtent()
# 
# xmin       : 71111.22 
# xmax       : 711789.6 
# ymin       : 4467237 
# ymax       : 4920994 

#e <- readRDS("../data/BGextent.rds")


# Convert spatial extent to spatial polygon

# library(Orcs)
# e <- ext2spy(e, crs = "EPSG:32635", as_sf= TRUE)

# Save extent polygon for later retrieval
# saveRDS(e, "../data/BGextent.rds")

# Get BG wider extent
e <- readRDS("../data/BGextent.rds")

# Plot all layers together (centered on BG)

plot(e);#plot(BG$geometry, col = "white", border = "black", lwd = 2, add = T); 
plot(BG$geometry, col = "white", border = "red", add = T); 
plot(Y_region$geometry,col = "black", border= "black", add = T);plot(GR$geometry, add = T);
plot(TR$geometry, add = T);plot(RO$geometry, add = T);plot(SE$geometry, add = T);plot(MK$geometry, add = T)
```


```{r inset-tmap}
# Make a map object with a neat style
bg_inset <- tm_shape(BG$geometry,bbox = e) + 
  tm_polygons(col = "white",
              border.col = "grey",
             lty = "solid",
             lwd = 2) + 
  tm_shape(GR$geometry) + 
  tm_polygons(col = "white",
              border.col = "grey",
             lty = "solid",
             lwd = 2) + 
  tm_shape(TR$geometry) + 
 tm_polygons(col = "white",
              border.col = "grey",
             lty = "solid",
             lwd = 2) + 
  tm_shape(RO$geometry) + 
  tm_polygons(col = "white",
              border.col = "grey",
             lty = "solid",
             lwd = 2) + 
  tm_shape(SE$geometry) + 
 tm_polygons(col = "white",
              border.col = "grey",
             lty = "solid",
             lwd = 2) + 
  tm_shape(MK$geometry) + 
 tm_polygons(col = "white",
              border.col = "grey",
             lty = "solid",
             lwd = 2) + 
  tm_shape(Y_region$geometry) + 
  tm_polygons(col = "grey",
             border.col = "grey" ,
             lty = "solid",
             lwd = 2) +
   tm_layout(bg.color="lightgrey")

bg_inset
```


Prepare to print with the inset. It prints to external file
```{r eval = FALSE}
library(tmaptools)
library(grid)

# change back to the plotting mode
#tmap_mode("plot")

pdf("../figures/Figure01.pdf", width = 5)
# PPAP
overview_yammnds

# print insets
print(bg_inset, vp=viewport(x= 0.17, y= 0.82, width= 0.3, height= 0.3))

dev.off()
```
