==> scripts/18_RSdatacheck2010.Rmd <==
---
title: "Enriching LU in mounds"
author: "Adela Sobotkova"
date: "23 June 2020"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## GOAL
The main aim of this script is to merge the topo_mounds 2010 dataset with data verified in Google Earth pro and gauge how much landuse information  can we extract from  RS data and what needs to be further verified with remote sensing.

Inputs:
1) topo_mounds which is a (full_join) merged product of adela and bara versions of 2010 data, counting some 289 individual records, where a mound was positively identified in the field (grepl("Burial Mound", Type) | SomethingPresentOntheGround == "mound"). Only Bara's records contain landuse, transcribed from the forms (which casts doubt on their quality, but theoretically has been verified in extant images as per LU_Source)
2) mounds_RS which is the product of remote sensing in GoogleEarth Pro of map features from 2010. It contains landuse cateories,which might help with topo_mounds


# QUICK CHECK: How much help is the RS dataset going to be?
```{r, echo = FALSE}
paste0("There are ",length(which(mounds_adela$TopoID%in%mounds_RS$TopoID)), " mounds shared between 2010 adela and RS dataset") 
paste0("There are ",length(which(mounds_bara$TopoID%in%mounds_RS$TopoID)), " mounds shared between 2010 bara and RS dataset") 
revLU_b <-  mounds_bara$TopoID[which(mounds_bara$TopoID%in%mounds_RS$TopoID)]
revLU_a <- mounds_adela$TopoID[which(mounds_adela$TopoID%in%mounds_RS$TopoID)]
revLU <- c(revLU_a, revLU_b)
paste0("There are ",length(revLU), " mounds shared between 2010 adela and bara, and RS dataset") 
paste0(length(unique(revLU))," shared mounds between both 2010 and RS datasets are unique.") 
```
RS dataset is not going to be very helpful for 2010 dataset. Maybe we should try it with 2017-2018 data instead.

# CHALLENGE: A NEW MERGE NEEDED and HERE WE NEED TO COMPARE BARA FIELD DATA WITH ADELA's RS DATA!
```{r}
topo_rs <- left_join(topo_mounds, mounds_RS, 
                         by = "TopoID", copy = FALSE, suffix = c(".topo", ".rs")) # 
```

And lets subset from the 46 columns to 16, picking only those columns of interest.

```{r}
topors <- topo_rs %>% 
  select(TopoID, Trap, TRAPCode, SomethingPresentOntheGround, Type, LandUseAround, LanduseOn, LU_Around, LU_top, Length.topo, `Length (max, m)`, Length.rs, RS_Status, RS_feature_type, Condition, Principal)

```

How many of the original 850 RS features have a counterpart in the Verified Mound data?

```{r}
topors %>%
  filter
  filter(!is.na(RS_Status))
```

It seems that only 24 mounds out of 850 possible mound features have a counterpart TopoID in the Verified mounds! That seems awfully low for 2010 season, although not impossible. (Maybe the filtering was too drastic previously?)
Consultation of the kmz files, the reason emerges:  remote sensig was done on "needGC" file, or the features we did not visit adequately to complement the 2010 field recording!
Anyways. The Landuse line of research: chceking the common discrepancies between fieldwalking and remote sensing classificaiton of landuse. >  see line XXX below,


Lets see how many LU problems can we solve or if the missingLU needs to be done in GoogleEarthPro. 

```{r}
topors %>% 
  filter(TopoID %in% missingLU$TopoID) %>% 
  select(TopoID, Trap, LU_Around, LU_top, Length.topo, `Length (max, m)`, Length.rs) %>% 
  filter(!is.na(LU_Around))
```

So the current data merge helps us solve 4 out of 34 missing LU (missing in Adela's dataset). That's not great.

## Merging fully
# just to verify I did not miss anything let's try to merge the other way around.
```{r}
rs_topo <- left_join(mounds_RS, topo_mounds, 
                         by = "TopoID", copy = FALSE, suffix = c(".rs", ".topo")) 
```

Now lets reduce the column number
```{r}
rstopo <- rs_topo %>% 
  select(TopoID, Trap, TRAPCode, SomethingPresentOntheGround, Type, LandUseAround, LanduseOn, LU_Around, LU_top, Length.topo, `Length (max, m)`, Length.rs, RS_Status, RS_feature_type, Condition, Principal)
```

And check RS status and number of attached mounds
```{r}
rstopo %>%
  filter(!is.na(RS_Status)) %>% 
  filter(grepl("Burial Mound", Type) | SomethingPresentOntheGround == "mound")
```

Still only 24 have some sort of moundlike status
Are the 34 mounds needing landuse in the RS dataset?

```{r}
which(missingLU$TopoID %in% rstopo$TopoID)
```


```{r}
missingLU %>% 
  filter(TopoID %in% rstopo$TopoID)
```

Compare this table with line 47. It checks out. So 4 Lu is all I get automatically from previous RS. Need to sense some 30 more mounds. Not a big deal and good to know. Maybe we can now bring in 2017 and 2018 data.


### Conclusion

RS data is not very helpful for 2010 datasets, perhaps because the RS spatial data was intended to complement the 2010 season not duplicate it. RS data should be helpful with 2017 -2018 seasons though - if they share a Topo id -  so merging with them is the next step. If TopoIDs don't exist, spatial join in ArcGIS will be a requisite.

Landuse in 2010 data (bara version) needs to be verified in GoogleEarth by exporting the kmx from ArcGIS > import to GEpro > review and then repeating the steps above. 

We need to see about 2017-2018 data next, before rushing to too dire conclusions..:)
